+++ demo/app/jobs/example_job.rb
@@ -19,7 +19,10 @@ class ExampleJob < ApplicationJob
     end
     def perform
       GoodJob::Batch.enqueue(on_finish: CallbackJob, description: "Example batch", foo: "bar") do
-        3.times { ExampleJob.perform_later(TYPES.sample) }
+        3.times do
+          job_type = TYPES.sample
+          ExampleJob.set(good_job_labels: [job_type]).perform_later(job_type)
+        end
       end
     end
   end
@@ -40,7 +43,26 @@ class ExampleJob < ApplicationJob
     elsif type == DEAD_TYPE
       raise DeadError
     elsif type == SLOW_TYPE
-      sleep 5
+      50.times do
+        break if blocking_reload? || GoodJob.current_thread_shutting_down?
+        sleep 0.1
+      end
+    end
+  end
+
+  private
+
+  def blocking_reload?
+    return false if Rails.application.config.cache_classes
+
+    ActiveSupport::Dependencies.interlock.raw_state do |threads|
+      # Find any thread attempting to unload (reload) code
+      return unless threads.any? { |_, info| info[:purpose] == :unload }
+
+      # Is the current thread blocking it? Likely yes.
+      info = threads[Thread.current]
+      info[:sharing] > 0 || # Thread holds a share lock
+        (info[:exclusive] && ![:load, :unload].include?(info[:purpose])) # Or holds incompatible exclusive lock
     end
   end
 end