+++ lib/good_job/configuration.rb
@@ -35,6 +35,8 @@ module GoodJob
     DEFAULT_DASHBOARD_LIVE_POLL_ENABLED = true
     # Default enqueue_after_transaction_commit
     DEFAULT_ENQUEUE_AFTER_TRANSACTION_COMMIT = false
+    # Default enable_pauses setting
+    DEFAULT_ENABLE_PAUSES = false
 
     def self.validate_execution_mode(execution_mode)
       raise ArgumentError, "GoodJob execution mode must be one of #{EXECUTION_MODES.join(', ')}. It was '#{execution_mode}' which is not valid." unless execution_mode.in?(EXECUTION_MODES)
@@ -369,6 +371,16 @@ module GoodJob
       DEFAULT_ENQUEUE_AFTER_TRANSACTION_COMMIT
     end
 
+    # Whether the job processing can be paused.
+    # @return [Boolean]
+    def enable_pauses
+      return options[:enable_pauses] unless options[:enable_pauses].nil?
+      return rails_config[:enable_pauses] unless rails_config[:enable_pauses].nil?
+      return ActiveModel::Type::Boolean.new.cast(env['GOOD_JOB_ENABLE_PAUSES']) unless env['GOOD_JOB_ENABLE_PAUSE'].nil?
+
+      DEFAULT_ENABLE_PAUSES
+    end
+
     # Whether running in a web server process.
     # @return [Boolean, nil]
     def in_webserver?
@@ -381,6 +393,14 @@ module GoodJob
           self_caller.grep(%{/rack/handler/}).any? || # EXAMPLE: iodine-0.7.44/lib/rack/handler/iodine.rb:13:in `start'
           (Concurrent.on_jruby? && self_caller.grep(%r{jruby/rack/rails_booter}).any?) # EXAMPLE: uri:classloader:/jruby/rack/rails_booter.rb:83:in `load_environment'
       end || false
+    end
+
+    def lower_thread_priority
+      return options[:lower_thread_priority] unless options[:lower_thread_priority].nil?
+      return rails_config[:lower_thread_priority] unless rails_config[:lower_thread_priority].nil?
+      return ActiveModel::Type::Boolean.new.cast(env['GOOD_JOB_LOWER_THREAD_PRIORITY']) unless env['GOOD_JOB_LOWER_THREAD_PRIORITY'].nil?
+
+      nil
     end
 
     # Whether to take an advisory lock on the process record in the notifier reactor.