+++ spec/lib/good_job/adapter_spec.rb
@@ -88,7 +88,7 @@ RSpec.describe GoodJob::Adapter do
         allow(GoodJob::Job).to receive(:enqueue).and_return(good_job)
         allow(GoodJob::Notifier).to receive(:notify)
 
-        capsule = instance_double(GoodJob::Capsule, start: nil, create_thread: nil)
+        capsule = instance_double(GoodJob::Capsule, start: nil, create_thread: nil, "lower_thread_priority=": nil)
         allow(GoodJob).to receive(:capsule).and_return(capsule)
         allow(capsule).to receive(:start)
 
@@ -98,6 +98,16 @@ RSpec.describe GoodJob::Adapter do
         expect(capsule).to have_received(:start)
         expect(capsule).to have_received(:create_thread)
         expect(GoodJob::Notifier).to have_received(:notify).with({ queue_name: 'default' })
+      end
+
+      it 'lowers the thread priority of the capsule' do
+        capsule = instance_double(GoodJob::Capsule, start: nil, create_thread: nil, "lower_thread_priority=": nil)
+        allow(GoodJob).to receive(:capsule).and_return(capsule)
+        allow(capsule).to receive(:start)
+
+        described_class.new(execution_mode: :async_all)
+
+        expect(capsule).to have_received(:lower_thread_priority=).with(true)
       end
     end
   end
@@ -155,6 +165,11 @@ RSpec.describe GoodJob::Adapter do
         priority: -55,
         scheduled_at: be_within(1).of(10.minutes.from_now)
       )
+      expect(GoodJob::Notifier).to have_received(:notify).with({
+                                                                 queue_name: 'elephant',
+                                                                 count: 1,
+                                                                 scheduled_at: within(1).of(10.minutes.from_now),
+                                                               })
     end
 
     it 'sets default values' do
@@ -166,6 +181,12 @@ RSpec.describe GoodJob::Adapter do
         priority: 0,
         scheduled_at: be_within(1.second).of(Time.current)
       )
+
+      expect(GoodJob::Notifier).to have_received(:notify).with({
+                                                                 queue_name: 'default',
+                                                                 count: 1,
+                                                                 scheduled_at: within(1).of(Time.current),
+                                                               })
     end
 
     context 'when a job fails to enqueue' do