+++ spec/lib/good_job_spec.rb
@@ -194,4 +194,23 @@ describe GoodJob do
       expect(described_class.deprecator).to have_received(:warn)
     end
   end
+
+  describe '.pause, .unpause, and .paused?' do
+    it 'can pause and unpause jobs' do
+      expect(described_class.paused?(queue: 'default')).to be false
+      expect(described_class.paused).to eq({ queues: [], job_classes: [], labels: [] })
+
+      described_class.pause(queue: 'default')
+      described_class.pause(job_class: 'MyJob')
+      expect(described_class.paused?(queue: 'default')).to be true
+      expect(described_class.paused).to eq({ queues: ['default'], job_classes: ['MyJob'], labels: [] })
+
+      described_class.unpause(queue: 'default')
+      expect(described_class.paused?(queue: 'default')).to be false
+      expect(described_class.paused).to eq({ queues: [], job_classes: ['MyJob'], labels: [] })
+
+      described_class.unpause(job_class: 'MyJob')
+      expect(described_class.paused).to eq({ queues: [], job_classes: [], labels: [] })
+    end
+  end
 end